local_system/apis/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-HTTPAuth==4.5.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
paho-mqtt==1.6.1
pkg-resources==0.0.0
typing-extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
